+++ benches/request_template_bench.rs
@@ -6,8 +6,8 @@ use hyper::HeaderMap;
 use serde_json::json;
 use tailcall::endpoint::Endpoint;
 use tailcall::has_headers::HasHeaders;
+use tailcall::http::RequestTemplate;
 use tailcall::path::PathString;
-use tailcall::request_template::RequestTemplate;
 
 #[derive(Setters)]
 struct Context {
