+++ src/blueprint/from_config/operators/http.rs
@@ -3,9 +3,8 @@ use crate::blueprint::*;
 use crate::config::group_by::GroupBy;
 use crate::config::{Config, Field};
 use crate::endpoint::Endpoint;
-use crate::http::Method;
+use crate::http::{Method, RequestTemplate};
 use crate::lambda::{Expression, Lambda, Unsafe};
-use crate::request_template::RequestTemplate;
 use crate::try_fold::TryFold;
 use crate::valid::{Valid, ValidationError};
 use crate::{config, helpers};
