+++ src/graphql/mod.rs
@@ -1,1 +1,5 @@
-pub mod dataloader;
+mod graphql_data_loader;
+mod graphql_request_template;
+
+pub use graphql_data_loader::*;
+pub use graphql_request_template::*;
