+++ src/http/mod.rs
@@ -2,12 +2,12 @@ mod client;
 mod data_loader;
 
 mod data_loader_request;
-mod graphql_data_loader;
 mod http_1;
 mod http_2;
 mod method;
 mod request_context;
 mod request_handler;
+mod request_template;
 mod response;
 mod server;
 mod server_config;
@@ -19,11 +19,11 @@ use cache_control::{Cachability, CacheControl};
 pub use client::*;
 pub use data_loader::*;
 pub use data_loader_request::*;
-pub use graphql_data_loader::*;
 use hyper::header::CACHE_CONTROL;
 pub use method::Method;
 pub use request_context::RequestContext;
 pub use request_handler::handle_request;
+pub use request_template::RequestTemplate;
 pub use response::*;
 pub use server::Server;
 pub use server_context::ServerContext;
