+++ src/lib.rs
@@ -4,11 +4,11 @@ pub mod blueprint;
 pub mod cache;
 pub mod cli;
 pub mod config;
+pub mod dataloader;
 pub mod directive;
 pub mod document;
 pub mod endpoint;
 pub mod graphql;
-pub mod graphql_request_template;
 pub mod has_headers;
 pub mod helpers;
 pub mod http;
@@ -19,6 +19,5 @@ pub mod lambda;
 pub mod mustache;
 pub mod path;
 pub mod print_schema;
-pub mod request_template;
 pub mod try_fold;
 pub mod valid;
